Send range data to future owner to reduce cache miss. (#1119)
After splitting, the split out ranges that locate on other nodes are PartiallyCached. When do query on these ranges, there will be a large number of cache misses.
In now design, we must scan out all records in the splitting range to determine the split position key (that is, all data are loaded into memory).
So, we can send the data locates in new ranges to their future owner nodes after data_sync_scan_op_ finishing. This can reduce the cache misses on new ranges with less costs.
Do not invalidate key cache on range split. Update the old range key cache during range split so that it can be used after range split. Initialize key cache on new range owner ng if it received the cache sent from old owner.

---------

+++ src/remote/cc_stream_receiver.cpp
@@ -1030,7 +1030,8 @@ void CcStreamReceiver::OnReceiveCcMsg(std::unique_ptr<CcMessage> msg)
                 for (int idx = 0; idx < cache_msg.scan_tuple_size(); ++idx)
                 {
                     const ScanTuple_msg &tuple_msg = cache_msg.scan_tuple(idx);
-                    assert(tuple_msg.cce_addr().core_id() == core_id);
+                    assert(tuple_msg.cce_addr().core_id() ==
+                           (uint32_t) core_id);
 
                     term = tuple_msg.cce_addr().term();
